Event price column datatype changed to Float

+++ backend/db/migrations/20221220005335-create-event.js
@@ -47,7 +47,7 @@ module.exports = {
 					allowNull: false,
 				},
 				price: {
-					type: Sequelize.DECIMAL(4, 2),
+					type: Sequelize.FLOAT,
 					allowNull: false,
 				},
 				startDate: {
